change expiration to duration

+++ powertools-idempotency/src/main/java/software/amazon/lambda/powertools/idempotency/persistence/BasePersistenceStore.java
@@ -52,7 +52,7 @@ public abstract class BasePersistenceStore implements PersistenceStore {
 
     private String functionName = "";
     private boolean configured = false;
-    private int expirationInSeconds = 60 * 60; // 1 hour default
+    private long expirationInSeconds = 60 * 60; // 1 hour default
     private boolean useLocalCache = false;
     private LRUCache<String, DataRecord> cache;
     private String eventKeyJMESPath;
